Move helper functions to `ipropertyvalue_helpers.dart`
